Moved likelihoods into their own files and cleaned up some files.
