+++ examples/MNIST_adam.py
@@ -8,6 +8,7 @@ import os
 import sys
 import matplotlib.pyplot as plt
 import torch.optim as optim
+from datetime import datetime
 from torch.utils.tensorboard import SummaryWriter
 
 from data_utils import read_data_sets
@@ -16,7 +17,7 @@ torch.set_default_dtype(torch.float32)
 
 sys.path.append("../src")
 
-from optim.FishLeg import FishLeg, FISH_LIKELIHOODS, initialise_FishModel
+from optim.FishLeg import FISH_LIKELIHOODS
 
 os.environ["CUDA_VISIBLE_DEVICES"] = "1"
 
@@ -62,48 +63,31 @@ model = nn.Sequential(
     nn.Linear(1000, 784, dtype=torch.float32),
 )
 
-model = initialise_FishModel(model, module_names="__ALL__")
-
 model = model.to(device)
 
 likelihood = FISH_LIKELIHOODS["bernoulli"](device=device)
 
-eta_adam = 1e-4
-
-lr = 0.02
-beta = 0.9
+lr = 0.001
+# betas = (0.7, 0.9)
 weight_decay = 1e-5
+# eps = 1e-8
 
-aux_lr = 1e-4
-aux_eps = 1e-8
-scale = 1
-damping = 0.5
-update_aux_every = 5
-
-initialization = "normal"
-normalization = True
-
-writer = SummaryWriter()
-
-opt = FishLeg(
-    model,
-    aux_loader,
-    likelihood,
-    lr=lr,
-    beta=beta,
-    weight_decay=weight_decay,
-    aux_lr=aux_lr,
-    aux_betas=(0.9, 0.999),
-    aux_eps=aux_eps,
-    fish_scale=scale,
-    damping=damping,
-    warmup_steps=1000,
-    update_aux_every=update_aux_every,
-    device=device,
-    writer=writer,
+writer = SummaryWriter(
+    log_dir=f"runs/MNIST_adam/lr={lr}_lambda={weight_decay}/{datetime.now().strftime('%Y%m%d-%H%M%S')}",
 )
 
-epochs = 10
+opt = optim.Adam(
+    model.parameters(),
+    lr=lr,
+    # betas=betas,
+    weight_decay=weight_decay,
+    # eps=eps,
+)
+
+epochs = 100
+
+st = time.time()
+eval_time = 0
 
 for epoch in range(1, epochs + 1):
     with tqdm(train_loader, unit="batch") as tepoch:
@@ -123,6 +107,7 @@ for epoch in range(1, epochs + 1):
             loss.backward()
             opt.step()
 
+            et = time.time()
             if n % 50 == 0:
                 model.eval()
 
@@ -143,7 +128,17 @@ for epoch in range(1, epochs + 1):
 
                 tepoch.set_postfix(loss=loss.item(), test_loss=running_test_loss)
                 model.train()
+                eval_time += time.time() - et
 
-        tepoch.set_postfix(loss=running_loss / n, test_loss=running_test_loss)
+        epoch_time = time.time() - st - eval_time
+
+        tepoch.set_postfix(
+            loss=running_loss / n, test_loss=running_test_loss, epoch_time=epoch_time
+        )
+        # Write out the losses per epoch
         writer.add_scalar("Loss/train", running_loss / n, epoch)
         writer.add_scalar("Loss/test", running_test_loss, epoch)
+
+        # Write out the losses per wall clock time
+        writer.add_scalar("Loss/train/time", running_loss / n, epoch_time)
+        writer.add_scalar("Loss/test/time", running_test_loss, epoch_time)
