+++ src/optim/FishLeg/__init__.py
@@ -1,11 +1,10 @@
 from .fishleg import FishLeg
 from .layers import *
-from .fishleg_likelihood import *
+from .likelihoods import *
 from .utils import *
 
 FISH_LAYERS = {"linear": FishLinear}
 FISH_LIKELIHOODS = {
-    "fixedgaussian": FixedGaussianLikelihood,
     "gaussian": GaussianLikelihood,
     "bernoulli": BernoulliLikelihood,
     "softmax": SoftMaxLikelihood,
