+++ src/optim/FishLeg/likelihoods/__init__.py
@@ -1,0 +1,4 @@
+from .likelihood_base import *
+from .bernoulli_likelihood import *
+from .gaussian_likelihood import *
+from .softmax_likelihood import *
